PDCS/Student/Rachita/requirements.txt
flask
pillow
tensorflow